LayoutTests/http/tests/filesystem/resolve-uri-expected.txt

Tests using resolveLocalFileSystemURI to obtain an Entry from a URI

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Resolving a generated URI.
PASS expectedPath is actualPath
PASS isFile is true
* Resolving test file by hand
PASS expectedPath is actualPath
PASS isFile is true
* Resolving a completely bogus URI.
PASS 
* Resolving a URI with the wrong protocol
PASS 
* Resolving a URI with no slash between type and file
PASS 
* Resolving a URI with no slash between protocol and type
PASS 
* Resolve a path using backslashes
PASS expectedPath is actualPath
PASS isFile is true
* Resolve a directory
PASS expectedPath is actualPath
PASS isDirectory is true
* Resolve a path using a trailing slash
PASS expectedPath is actualPath
PASS isDirectory is true
* Resolving a persistent URI.
PASS expectedPath is actualPath
PASS isDirectory is true
Finished running tests.
PASS successfullyParsed is true

TEST COMPLETE
